combine duplicate interpolation code from mercator and globe
